+++ instrumentation/grpc-1.5/src/main/java/io/opentelemetry/instrumentation/auto/grpc/server/TracingServerInterceptor.java
@@ -34,7 +34,7 @@ import io.opentelemetry.trace.Span;
 import io.opentelemetry.trace.attributes.SemanticAttributes;
 import java.net.InetSocketAddress;
 import java.net.SocketAddress;
-import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicLong;
 
 public class TracingServerInterceptor implements ServerInterceptor {
 
@@ -97,7 +97,7 @@ public class TracingServerInterceptor implements ServerInterceptor {
   static final class TracingServerCallListener<ReqT>
       extends ForwardingServerCallListener.SimpleForwardingServerCallListener<ReqT> {
     private final Span span;
-    private final AtomicInteger messageId = new AtomicInteger();
+    private final AtomicLong messageId = new AtomicLong();
 
     TracingServerCallListener(Span span, ServerCall.Listener<ReqT> delegate) {
       super(delegate);
@@ -111,7 +111,7 @@ public class TracingServerInterceptor implements ServerInterceptor {
               SemanticAttributes.GRPC_MESSAGE_TYPE,
               "RECEIVED",
               SemanticAttributes.GRPC_MESSAGE_ID,
-              (long) messageId.incrementAndGet());
+              messageId.incrementAndGet());
       span.addEvent("message", attributes);
       try (Scope ignored = currentContextWith(span)) {
         delegate().onMessage(message);
